updated get_star() function , UI and Ko-fi links
